pyproject.toml
[project]
name = "stt-cli"
version = "0.1.0"
description = "Speech-to-text CLI with Google Vertex AI, supporting speaker diarization and automatic language detection"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "google-cloud-speech>=2.20.0",
    "google-cloud-storage>=2.10.0",
]

[project.scripts]
stt-cli = "stt_cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
]
